adding speed.py, correct indentation problem on distance.py
