+++ distance.py
@@ -1,4 +1,4 @@
-  
+
 '''
 -------------------------------------------
 -    Author: Asadullah Dal                -
@@ -11,83 +11,92 @@
 -------------------------------------------
 '''
 
-import cv2 
-# variables 
+import cv2
+# variables
 # distance from camera to object(face) measured
-Known_distance = 76.2 #centimeter
+Known_distance = 76.2  # centimeter
 # width of face in the real world or Object Plane
-Known_width =14.3 #centimeter
+Known_width = 14.3  # centimeter
 # Colors
-GREEN = (0,255,0)
-RED = (0,0,255)
-WHITE = (255,255,255)
+GREEN = (0, 255, 0)
+RED = (0, 0, 255)
+WHITE = (255, 255, 255)
 fonts = cv2.FONT_HERSHEY_COMPLEX
 cap = cv2.VideoCapture(1)
 # face detector object
 face_detector = cv2.CascadeClassifier("haarcascade_frontalface_default.xml")
-# focal length finder function  
+# focal length finder function
+
+
 def FocalLength(measured_distance, real_width, width_in_rf_image):
-	'''
-    This Function Calculate the Focal Length(distance between lens to CMOS sensor), it is simple constant we can find by using 
-    MEASURED_DISTACE, REAL_WIDTH(Actual width of object) and WIDTH_OF_OBJECT_IN_IMAGE 
-    :param1 Measure_Distance(int): It is distance measured from object to the Camera while Capturing Reference image
-    
-    :param2 Real_Width(int): It is Actual width of object, in real world (like My face width is = 14.3 centimeters)
-    :param3 Width_In_Image(int): It is object width in the frame /image in our case in the reference image(found by Face detector) 
-    :retrun Focal_Length(Float):
     '''
-	focal_length = (width_in_rf_image* measured_distance)/ real_width
-	return focal_length
+This Function Calculate the Focal Length(distance between lens to CMOS sensor), it is simple constant we can find by using 
+MEASURED_DISTACE, REAL_WIDTH(Actual width of object) and WIDTH_OF_OBJECT_IN_IMAGE 
+:param1 Measure_Distance(int): It is distance measured from object to the Camera while Capturing Reference image
+
+:param2 Real_Width(int): It is Actual width of object, in real world (like My face width is = 14.3 centimeters)
+:param3 Width_In_Image(int): It is object width in the frame /image in our case in the reference image(found by Face detector) 
+:retrun Focal_Length(Float):
+'''
+    focal_length = (width_in_rf_image * measured_distance) / real_width
+    return focal_length
 # distance estimation function
-def Distance_finder (Focal_Length, real_face_width, face_width_in_frame):
-		'''
-    This Function simply Estimates the distance between object and camera using arguments(Focal_Length, Actual_object_width, Object_width_in_the_image)
-    :param1 Focal_length(float): return by the Focal_Length_Finder function
-    
-    :param2 Real_Width(int): It is Actual width of object, in real world (like My face width is = 5.7 Inches)
-    :param3 object_Width_Frame(int): width of object in the image(frame in our case, using Video feed)
-    :return Distance(float) : distance Estimated  
-    
+
+
+def Distance_finder(Focal_Length, real_face_width, face_width_in_frame):
     '''
-	distance = (real_face_width * Focal_Length)/face_width_in_frame
-	return distance
+This Function simply Estimates the distance between object and camera using arguments(Focal_Length, Actual_object_width, Object_width_in_the_image)
+:param1 Focal_length(float): return by the Focal_Length_Finder function
+
+:param2 Real_Width(int): It is Actual width of object, in real world (like My face width is = 5.7 Inches)
+:param3 object_Width_Frame(int): width of object in the image(frame in our case, using Video feed)
+:return Distance(float) : distance Estimated  
+
+'''
+    distance = (real_face_width * Focal_Length)/face_width_in_frame
+    return distance
+
 
 def face_data(image):
-	'''
-	This function Detect the face 
-	:param Takes image as argument.
-	:returns face_width in the pixels
-	'''
+    '''
+    This function Detect the face 
+    :param Takes image as argument.
+    :returns face_width in the pixels
+    '''
 
-	face_width = 0 
-	gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
-	faces = face_detector.detectMultiScale(gray_image, 1.3, 5)
-	for (x, y, h, w) in faces:
-		cv2.rectangle(image, (x, y), (x+w, y+h), WHITE, 1)
-		face_width = w
+    face_width = 0
+    gray_image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
+    faces = face_detector.detectMultiScale(gray_image, 1.3, 5)
+    for (x, y, h, w) in faces:
+        cv2.rectangle(image, (x, y), (x+w, y+h), WHITE, 1)
+        face_width = w
 
-	return face_width
+    return face_width
+
 
 # reading reference image from directory
 ref_image = cv2.imread("Ref_image.png")
 
 ref_image_face_width = face_data(ref_image)
-Focal_length_found = FocalLength(Known_distance, Known_width, ref_image_face_width)
+Focal_length_found = FocalLength(
+    Known_distance, Known_width, ref_image_face_width)
 print(Focal_length_found)
 cv2.imshow("ref_image", ref_image)
 
 while True:
-	_, frame = cap.read()
+    _, frame = cap.read()
 
-	# calling face_data function
-	face_width_in_frame = face_data(frame)
-	# finding the distance by calling function Distance 
-	if face_width_in_frame !=0:
-		Distance = Distance_finder(Focal_length_found, Known_width,face_width_in_frame)
-	# Drwaing Text on the screen
-		cv2.putText(frame, f"Distance = {round(Distance,2)} CM", (50,50), fonts,1, (WHITE),2)
-	cv2.imshow("frame", frame )
-	if cv2.waitKey(1)==ord("q"):
-		break 
+    # calling face_data function
+    face_width_in_frame = face_data(frame)
+    # finding the distance by calling function Distance
+    if face_width_in_frame != 0:
+        Distance = Distance_finder(
+            Focal_length_found, Known_width, face_width_in_frame)
+    # Drwaing Text on the screen
+        cv2.putText(
+            frame, f"Distance = {round(Distance,2)} CM", (50, 50), fonts, 1, (WHITE), 2)
+    cv2.imshow("frame", frame)
+    if cv2.waitKey(1) == ord("q"):
+        break
 cap.release()
 cv2.destroyAllWindows()
